refactor
- relocate some files
- replace some variables
