+++ lib/core/constants.dart
@@ -1,0 +1,1 @@
+const kZenkakuGothicNew = 'ZenkakuGothicNew';
